config.json: Mount a tmpfs to /var/lib/nginx/tmp
Avoid:

  $ ccon
  nginx: [emerg] mkdir() "/var/lib/nginx/tmp/client" failed (2: No such file or directory)

That's the client_body_temp_path [1].  Using a tmpfs increases the
denial-of-service exposure, because you probably have more disk space
than you have memory, but the post-pivot root filesystem is read only
so we need something there.  An alternative would be to bind in a
writable directory from the host, but for now I prefer the automatic
cleanup of a tmpfs.

[1]: https://nginx.org/en/docs/http/ngx_http_core_module.html#client_body_temp_path