root/etc/nginx/nginx.conf: Clobber Vary
The upstream server is setting User-Agent:

  $ curl -ksI --resolve hl.azurestandard.com:443:192.168.161.1
  >   https://hl.azurestandard.com/healthy-living/ |
  >   grep -i '^http\|vary'
  HTTP/1.1 200 OK
  Vary: Accept-Encoding,Cookie,User-Agent

I'm not clear on why we'd want User-Agent; it seems like it would make
caching much less efficient.  I'm clobbering it away with this commit.

We want to set Origin because we use that value in
Access-Control-Allow-Origin (the wildcard * is not allowed for
requests that support credentials [1]).

[1]: https://www.w3.org/TR/2014/REC-cors-20140116/#resource-requests