root/etc/nginx/nginx.conf: Shift ssl_* configuration to the http block
We usually have a number of subdomain servers all sharing the same
wildcard cert.  By moving these settings to the http block we don't
have to repeat them for each server.  And you can still clobber them
per-server if you have an instance or two that need a different
certificate.